Refactor TaskContext: Extract types, improve error handling, and cleanup
