+++ src/hooks/useActivity.ts
@@ -30,7 +30,7 @@ export interface Activity {
   userAvatar?: string;
   targetUserId?: string; // For assignments
   targetUserName?: string;
-  metadata?: Record<string, any>;
+  metadata?: Record<string, unknown>;
   createdAt: Timestamp;
 }
 
@@ -40,7 +40,7 @@ export interface CreateActivityInput {
   groupId: string;
   targetUserId?: string;
   targetUserName?: string;
-  metadata?: Record<string, any>;
+  metadata?: Record<string, unknown>;
 }
 
 export interface UseActivityOptions {
@@ -130,7 +130,7 @@ export const useActivity = (
             setActivities(activityList);
             setLoading(false);
           },
-          err => {
+          _err => {
             setError('활동 정보를 불러오는 중 오류가 발생했습니다.');
             setLoading(false);
           }
@@ -138,7 +138,7 @@ export const useActivity = (
 
         return unsubscribe;
       }
-    } catch (err) {
+    } catch (_err) {
       setError('활동 정보를 불러오는 중 오류가 발생했습니다.');
       setLoading(false);
     }
@@ -158,7 +158,7 @@ export const useActivity = (
         };
 
         await addDoc(collection(db, 'activities'), activityDoc);
-      } catch (err) {
+      } catch (_err) {
         // Don't throw error to prevent disrupting main functionality
       }
     },
@@ -223,7 +223,7 @@ export const usePresence = (groupId: string): UsePresenceReturn => {
         };
 
         await addDoc(collection(db, 'presence'), presenceDoc);
-      } catch (error) {
+      } catch (_error) {
         // Handle error silently
       }
     },
